# api-spec.yml
openapi: 3.0.0
info:
  title: User Bank Details API
  description: API for managing CRUD operations for user bank accounts.
  version: 1.1.0

paths:
  # --- AUTHENTICATION ENDPOINTS ---
  /auth/signup:
    post:
      tags: [ Authentication ]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignupRequest'
      responses:
        '201':
          description: User registered successfully
        '400':
          description: Username or Email already exists

  /auth/login:
    post:
      tags: [ Authentication ]
      summary: Authenticate user and return JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials

  /auth/signout:
    post:
      tags: [ Authentication ]
      summary: Sign out the current user
      security:
        - BearerAuth: [ ]
      responses:
        '200':
          description: Successfully signed out
        '401':
          description: Unauthorized (Token missing or expired)
  /bankDetails:
    get:
      summary: Get all bank accounts for a user
      description: Returns a list of all bank accounts associated with a specific userId.
      parameters:
        - name: userId
          in: query
          required: true
          description: The ID of the user whose accounts you want to retrieve.
          schema:
            type: string
          example: "user_98765"
      responses:
        '200':
          description: A list of bank accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BankDetailResponse'
        '400':
          description: Missing userId parameter
    post:
      summary: Create a new bank entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BankDetail'
      responses:
        '201':
          description: Bank detail created successfully

  /bank-details/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      summary: Get a specific bank record by ID
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BankDetailResponse'
    put:
      summary: Update bank details
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BankDetail'
      responses:
        '200':
          description: Updated successfully
    delete:
      summary: Remove a bank entry
      responses:
        '204':
          description: Deleted successfully

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    SignupRequest:
        type: object
        required: [ username, emailId, password ]
        properties:
          username:
            type: string
            example: "johndoe"
          emailId:
            type: string
            format: email
            example: "john@example.com"
          password:
            type: string
            format: password
            example: "P@ssw0rd123"

    LoginRequest:
        type: object
        required: [ username, password ]
        properties:
          username:
            type: string
            example: "johndoe"
          password:
            type: string
            format: password
            example: "P@ssw0rd123"

    AuthResponse:
        type: object
        properties:
          token:
            type: string
            description: "The JWT token to be used in the Authorization header"
          type:
            type: string
            example: "Bearer"
          username:
            type: string
          userId:
            type: string

    BankDetail:
      type: object
      required: [userId, accountName, accountNumber, bankName, IFSC_Code]
      properties:
        userId:
          type: string
        accountName:
          type: string
        accountNumber:
          type: string
        bankName:
          type: string
        IFSC_Code:
          type: string
    BankDetailResponse:
      type: object
      properties:
        id:
          type: string
          description: Internal database record ID
        userId:
          type: string
        accountName:
          type: string
        accountNumber:
          type: string
        bankName:
          type: string
        IFSC_Code:
          type: string